+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Changed
+- Updates to how we apply colors to `ButtonOutline` and `ButtonTransparent` to be more inline with design spec
+- `IconButton` now uses ui color by default
+- `neutral` intent color is now defaults to `charcoal500`
+
 ## [0.9.7] - 2020-07-27
 
 ### Added